Branch Structure 	- Main.cs 		- Fixed comments 	- Singleton.cs 		- Added comments 	- BaseSingleton.cs 		- Moved BaseSingleton to it's own cs file 		- Added comments 	- Jobs.cs 		- Moved Jobs to it's own cs file 		- Added comments 	- Reps.cs 		- Moved Reps to it's own cs file 		- Added comments 	- Companies.cs 		- Moved Companies to it's own cs file 		- Added comments 	- Enums.cs 		- Added comments 	- BaseListWindow.xaml.cs 		- Removed unused using 	-BaseCreationWindow.xaml.cs 		- Removed unused using 	- CompanyCreationWindow.xaml.cs 		- Removed unused using 	- Company.cs 		- Removed unused using 	- CompanyCreation.cs 		- Removed unused using 	- Commands.cs 		- Added comments
