+++ BaseCreationWindow.xaml.cs
@@ -1,16 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
+﻿using System.Windows;
 
 namespace Job_Application_Database
 {
@@ -27,4 +15,5 @@ namespace Job_Application_Database
             Exit = Enum.ExitStatus.Cancel;
         }
     }
+
 }
