+++ BaseListWindow.xaml.cs
@@ -1,20 +1,4 @@
-﻿using Job_Application_Database.Classes;
-using System;
-using System.ComponentModel;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
-using Job_Application_Database.IO;
-using Job_Application_Database.Singleton;
+﻿using System.Windows;
 
 namespace Job_Application_Database
 {
@@ -32,4 +16,5 @@ namespace Job_Application_Database
 
         }
     }
+
 }
