+++ BaseSingleton.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Job_Application_Database.Singleton
+{
+    /// <summary>
+    /// Base Singleton Class To Make Common Inheritance
+    /// </summary>
+    public class BaseSingleton : Singleton<BaseSingleton>
+    {
+        // Intentionally Left Blank
+    }
+
+}
