+++ CompanyCreationWindow.xaml.cs
@@ -1,10 +1,4 @@
-﻿using Job_Application_Database.Classes;
-using Job_Application_Database.Singleton;
-using System;
-using System.ComponentModel;
-using System.Windows;
-using System.Windows.Data;
-using System.Windows.Input;
+﻿using System.Windows;
 
 namespace Job_Application_Database
 {
@@ -23,4 +17,5 @@ namespace Job_Application_Database
             Exit = Enum.ExitStatus.Cancel;
         }
     }
+
 }
