+++ Singletons.cs
@@ -1,24 +1,28 @@
 ﻿using Job_Application_Database.Classes;
-using Newtonsoft.Json;
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Job_Application_Database.Singleton
 {
-
+    /// <summary>
+    /// Singleton Class
+    /// </summary>
     public abstract class Singleton<T> where T : class, new()
     {
-
+        // The Lazy Construction Of The Instance
         private static readonly Lazy<T> _instance = new Lazy<T>(() => CreateInstance());
 
+        // To Tell If _allobjectnames Needs To Be Refreshed
         protected bool _needsRefresh = false;
+
+        // Hash Table To Store All BaseInfo
         protected Hashtable _table = new Hashtable();
+
+        // List To Store All BaseInfo Names
         protected List<string> _allobjectnames;
 
+        // The Instance Of The Singleton
         public static T Instance
         {
             get
@@ -27,18 +31,20 @@ namespace Job_Application_Database.Singleton
             }
         }
 
+        // Creation Of The Instance
         private static T CreateInstance()
         {
             return Activator.CreateInstance(typeof(T), true) as T;
         }
 
+        // Default Constructor
         protected Singleton()
         {
             _table = new Hashtable();
             _allobjectnames = new List<string>();
         }
 
-
+        // The Size Of The Hash Table
         public int Count
         {
             get
@@ -47,8 +53,10 @@ namespace Job_Application_Database.Singleton
             }
         }
 
+        // Used To Initialized The Hash Table With BaseInfo
         public virtual void InitObjectList() { }
 
+        // Refreshes The Names List
         protected void Refresh()
         {
             _allobjectnames.Clear();
@@ -61,6 +69,7 @@ namespace Job_Application_Database.Singleton
             _needsRefresh = false;
         }
 
+        // Gets All Of The BaseInfo As A List
         public List<BaseInfo> AllObjects()
         {
             List<BaseInfo> list = new List<BaseInfo>();
@@ -77,29 +86,34 @@ namespace Job_Application_Database.Singleton
             return list;
         }
 
+        // Adds BaseInfo To The Hash Table
         public void AddObject(BaseInfo k)
         {
             _table.Add(k.ID, k);
             _needsRefresh = true;
         }
 
+        // Removes BaseInfo From Hast Table
         public void RemoveObject(BaseInfo k)
         {
             _table.Remove(k.ID);
             _needsRefresh = true;
         }
 
+        // Gets The BaseInfo With Spesified ID 
         public BaseInfo ObjectAt(int id)
         {
             return _table[id] as BaseInfo;
         }
 
+        // Gets All BaseInfo Names As A List
         public List<string> AllObjectNames()
         {
             if (_needsRefresh) Refresh();
             return _allobjectnames;
         }
 
+        // Gets The Location Of BaseInfo In _allobjectnames From It's ID
         public int TableToNames(int id)
         {
             Refresh();
@@ -113,6 +127,7 @@ namespace Job_Application_Database.Singleton
             return -1;
         }
 
+        // Gets The BaseInfo In _table From It's Location in _allobjectnames
         public BaseInfo NamesToTable(int index)
         {
             string name = _allobjectnames[index];
@@ -128,94 +143,4 @@ namespace Job_Application_Database.Singleton
         }
     }
 
-    public class BaseSingleton : Singleton<BaseSingleton>
-    {
-        
-    }
-
-    public sealed class Jobs : BaseSingleton
-    {
-        public new static Jobs Instance
-        {
-            get
-            {
-                return Singleton<Jobs>.Instance;
-            }
-        }
-
-        public override void InitObjectList()
-        {
-            _table = new Hashtable
-            {
-                { 0, new Job() },
-                { 1, new Job("Android Developer") },
-                { 2, new Job("Embedded Software Engineer I") },
-                { 3, new Job("Full Stack Developer I") },
-                { 4, new Job("Game Developer") },
-                { 5, new Job("Game R&D") },
-                { 6, new Job("Java Developer") },
-                { 7, new Job("Mobile Game Developer") },
-                { 8, new Job("Software Engineer I") },
-                { 9, new Job("Software Tester I") }
-            };
-
-            _needsRefresh = true;
-        }
-    }
-
-    public sealed class Reps : BaseSingleton
-    {
-
-        public new static Reps Instance
-        {
-            get
-            {
-                return Singleton<Reps>.Instance;
-            }
-        }
-
-        public override void InitObjectList()
-        {
-            _table = new Hashtable
-            {
-                { 0, new Rep() }
-            };
-
-            _needsRefresh = true;
-        }
-    }
-
-    public sealed class Companies : BaseSingleton
-    {
-        public new static Companies Instance
-        {
-            get
-            {
-                return Singleton<Companies>.Instance;
-            }
-        }
-    }
-
-    public sealed class Boards : BaseSingleton
-    {
-        public new static Boards Instance
-        {
-            get
-            {
-                return Singleton<Boards>.Instance;
-            }
-        }
-
-        public override void InitObjectList()
-        {
-            _table = new Hashtable
-            {
-                { 0, new Board(" ") },
-                { 1, new Board("Glassdoor", "www.glassdoor.com") },
-                { 2, new Board("Indeed", "www.indeed.com") },
-                { 3, new Board("Linkedin", "www.linkedin.com") },
-                { 4, new Board("Monster", "www.monster.com") }
-            };
-        }
-    }
 }
