+++ bin/Debug/job.titles
@@ -3,6 +3,7 @@
 {"Title":"Automation Engineer","Salary":null,"ID":11}
 {"Title":"Backend Software Engineer","Salary":"~","ID":12}
 {"Title":"Build Engineer","Salary":"~","ID":16}
+{"Title":"C++ Embedded Developer","Salary":"~","ID":17}
 {"Title":"Embedded Software Engineer I","Salary":null,"ID":2}
 {"Title":"Full Stack Developer I","Salary":null,"ID":3}
 {"Title":"Game Developer","Salary":null,"ID":4}
